give money

+++ wechat-master/src/main/java/fr/pantheonsorbonne/ufr27/miage/model/Help.java
@@ -14,17 +14,17 @@ public class Help {
     private String region;
 
     @Column(name = "moneySupport", nullable = false)
-    private int moneySupport;
+    private double moneySupport;
 
     @Column(name = "timeSupport", nullable = false)
-    private int timeSupport;
+    private double timeSupport;
 
     @Column(name = "clotheSupport", nullable = false)
-    private int clotheSupport;
+    private double clotheSupport;
 
 
 
-    public Help(int helpId, String region, int moneySupport, int timeSupport, int clotheSupport) {
+    public Help(int helpId, String region, double moneySupport, double timeSupport, double clotheSupport) {
         this.helpId  = helpId;
         this.region=region;
         this.moneySupport = moneySupport;
@@ -51,27 +51,27 @@ public class Help {
         this.region = region;
     }
 
-    public Integer getMoneySupport() {
+    public double getMoneySupport() {
         return moneySupport;
     }
 
-    public void setMoneySupport(Integer moneySupport) {
+    public void setMoneySupport(double moneySupport) {
         this.moneySupport = moneySupport;
     }
 
-    public Integer getTimeSupport() {
+    public double getTimeSupport() {
         return timeSupport;
     }
 
-    public void setTimeSupport(Integer timeSupport) {
+    public void setTimeSupport(double timeSupport) {
         this.timeSupport = timeSupport;
     }
 
-    public Integer getClotheSupport() {
+    public double getClotheSupport() {
         return clotheSupport;
     }
 
-    public void setClotheSupport(Integer clotheSupport) {
+    public void setClotheSupport(double clotheSupport) {
         this.clotheSupport = clotheSupport;
     }
